referai-ai-service/pyproject.toml
[project]
name = "referai-ai-service"
version = "0.2.0"
description = "ReferAI Python AI microservice — LangGraph agents, semantic matching, coaching"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.32.1",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.1",
    "asyncpg>=0.30.0",
    "redis>=5.2.1",
    "structlog>=24.4.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "httpx>=0.27.0",
    "openai>=1.52.0",
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "google-generativeai>=0.8.0",
    "pgvector>=0.3.0",
    "sse-starlette>=2.0.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
]

[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["migrations*", "tests*"]
